League of Legends DataSet
This dataset is from a popular MOBA game called League of Legends, a tower deffense game. With 40 variables
and thousands of entries the goal was to predict the outcome of a match depending on the variable values
at the first 10 min of a match.

A game of League of Legends is team game where to team compete to conquer the opposite side. Each team has 
5 players. Wins the team that destroys the opposite nexus bringing the match to an end.

both teams have access to wards (that grant vision in dark areas of the map), gold score, experience score,
level score, kill, death, assist, creep score. And these are the variables used to make predictions on the
outcome.